+++ spec/lib/karafka/web/pro/ui/controllers/errors_controller_spec.rb
@@ -249,11 +249,43 @@ RSpec.describe_current do
       it do
         expect(response).to be_ok
         expect(body).to include('shinra:1555833:4e8f7174ae53')
-        expect(body.scan('StandardError').size).to eq(4)
+        expect(body.scan('StandardError').size).to eq(3)
         expect(body).to include(breadcrumbs)
         expect(body).not_to include(pagination)
         expect(body).not_to include(support_message)
         expect(body).not_to include('This feature is available only')
+      end
+    end
+
+    context 'when visiting error offset with a transactional record in range' do
+      before do
+        produce(errors_topic, error_report, partition: 0, type: :transactional)
+
+        get 'errors/0/1'
+      end
+
+      it do
+        expect(response).to be_ok
+        expect(body).not_to include('shinra:1555833:4e8f7174ae53')
+        expect(body).not_to include('StandardError')
+        expect(body).to include(breadcrumbs)
+        expect(body).to include(pagination)
+        expect(body).to include('The message has been removed')
+        expect(body).not_to include(support_message)
+        expect(body).not_to include('This feature is available only')
+      end
+    end
+
+    context 'when visiting offset on transactional above watermark' do
+      before do
+        produce(errors_topic, error_report, partition: 0, type: :transactional)
+
+        get 'errors/0/2'
+      end
+
+      it do
+        expect(response).not_to be_ok
+        expect(status).to eq(404)
       end
     end